fix files names and migrations in sql
